requirements.txt
aiodns==3.0.0
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pkg_resources==0.0.0
pycares==4.1.2
pycparser==2.21
python-dotenv==0.19.2
pytz==2021.3
typing_extensions==4.0.0
ujson==4.3.0
uvloop==0.16.0
yarl==1.7.2
